@@ -13,20 +13,20 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "0.0.34",
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.4",
-    "angular-4-data-table": "^0.2.3",
-    "angular4-paystack": "^1.2.1",
-    "angularfire2": "^4.0.0-rc0",
+    "@angular-devkit/schematics": "7.0.3",
+    "@angular/animations": "^7.0.1",
+    "@angular/common": "^7.0.1",
+    "@angular/compiler": "^7.0.1",
+    "@angular/core": "^7.0.1",
+    "@angular/forms": "^7.0.1",
+    "@angular/http": "^7.0.1",
+    "@angular/platform-browser": "^7.0.1",
+    "@angular/platform-browser-dynamic": "^7.0.1",
+    "@angular/router": "^7.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "angular4-paystack": "^2.2.0",
+    "angular5-data-table": "^1.0.3",
+    "angularfire2": "^5.1.0",
     "bootstrap": "^4.0.0-beta",
     "core-js": "^2.4.1",
     "firebase": "^5.5.6",
@@ -38,28 +38,30 @@
     "ngx-carousel": "^1.3.2",
     "ngx-owl-carousel": "^2.0.1",
     "popper.js": "^1.11.0",
-    "rxjs": "^5.4.2",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
+    "@angular-devkit/build-angular": "~0.10.0",
     "@angular/cli": "^7.0.3",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/jasmine": "~2.5.53",
+    "@angular/compiler-cli": "^7.0.1",
+    "@angular/language-service": "^7.0.1",
+    "@types/jasmine": "~2.8.9",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
+    "@types/node": "~10.12.0",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.3.0",
+    "jasmine-spec-reporter": "~4.2.1",
     "karma": "^3.1.1",
-    "karma-chrome-launcher": "~2.1.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "^5.4.1",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "^2.3.4"
+    "ts-node": "~7.0.1",
+    "tslint": "~5.11.0",
+    "typescript": "^3.1.3"
   }
 }
