@@ -16,7 +16,7 @@ import { AngularFireModule } from 'angularfire2';
 import { AngularFireDatabaseModule } from 'angularfire2/database';
 import { AngularFireAuthModule } from 'angularfire2/auth';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
-import { DataTableModule } from 'angular-4-data-table';
+import { DataTableModule } from 'angular5-data-table';
 import { Angular4PaystackModule } from 'angular4-paystack';
 
 import { AppComponent } from './app.component';
