@@ -1,4 +1,4 @@
-export interface AppUser {
+export interface AppUser extends firebase.User {
     name: string;
     email: string;
     isAdmin: boolean;
